+++ tests/reorder_code/expected/reorder_region_definition_spacing.gd
@@ -1,16 +1,4 @@
-# #region / #endregion markers
-# Region start and end comments stay next to the lines they enclose, but spacing
-# is applied around the region based on the enclosed content.
-#region Description
-func test():
-	pass
-#endregion
-
-
-#region AnotherRegion
-func test2():
-	pass
-#endregion
+extends Node
 
 
 func foo() -> void:
@@ -29,8 +17,7 @@ func after_region():
 
 #region B (outer region)
 #region C (nested inside B)
-
-## Test docstring
+## A documented function.
 @warning_ignore("unused_parameter")
 func nested(value):
 	pass